feat(folders-list-container): add delete folder dialog confirmation

+++ libs/users/materials/feature-folders-list/src/lib/folders-list-container/folders-list-container.component.ts
@@ -1,10 +1,14 @@
 import { CommonModule } from '@angular/common';
-import { ChangeDetectionStrategy, Component, inject, ViewEncapsulation } from '@angular/core';
+import { ChangeDetectionStrategy, Component, DestroyRef, inject, ViewEncapsulation } from '@angular/core';
 import { LetDirective } from '@ngrx/component';
 import { FoldersAddButtonComponent } from '@users/feature-folders-create';
 import { MaterialsFacade } from '@users/materials/data-access';
 import { FoldersListComponent } from '../folders-list/folders-list.component';
 import { IFolder } from 'libs/users/materials/data-access/src/lib/models/folder.model';
+import { CoreUiConfirmDialogComponent } from '@users/core/ui';
+import { MatDialog, MatDialogRef } from '@angular/material/dialog';
+import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'users-folders-list-container',
@@ -18,12 +22,29 @@ import { IFolder } from 'libs/users/materials/data-access/src/lib/models/folder.
 export class FoldersListContainerComponent {
   private readonly materialsFacade = inject(MaterialsFacade);
   public folders$ = this.materialsFacade.folders$;
+  private readonly dialog = inject(MatDialog);
+  private readonly destroyRef = inject(DestroyRef);
+  private readonly router = inject(Router);
 
   ngOnInit(): void {
     this.materialsFacade.init();
   }
 
   onDeleteFolder(folder: IFolder) {
-    this.materialsFacade.deleteFolder(folder.id);
+    const dialogRef: MatDialogRef<CoreUiConfirmDialogComponent> = this.dialog.open(CoreUiConfirmDialogComponent, {
+      data: {
+        dialogText: `Вы уверены, что хотите удалить папку ${folder.title}?`,
+      },
+    });
+
+    dialogRef
+      .afterClosed()
+      .pipe(takeUntilDestroyed(this.destroyRef))
+      .subscribe((result) => {
+        if (result) {
+          this.materialsFacade.deleteFolder(folder.id);
+          this.router.navigate(['/materials']);
+        }
+      });
   }
 }
